icu 58.2 (release), original sources
git-svn-id: svn://tug.org/texlive/trunk/Build/source@43491 c570f23f-e606-0410-a88d-b1316a301751

+++ libs/icu/icu-src/source/test/cintltst/bocu1tst.c
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 ******************************************************************************
 *
@@ -342,7 +344,7 @@ packDiff(int32_t diff) {
     } while(--count>0);
 
     /* add lead byte */
-    result|=(lead+diff)<<shift;
+    result |= (uint32_t)(lead+diff)<<shift;
 
     return result;
 }
